向事实数据表添加数据定时任务

+++ core/src/main/java/com/tianqiauto/textile/weaving/model/sys/Current_BuJi.java
@@ -79,7 +79,7 @@ public class Current_BuJi {
 
     private Double weimi; //纬密 根/inch
 
-    private Double daweicishu; //当班打纬次数
+    private Long daweicishu; //当班打纬次数
 
     private Double suolv; //缩率
